Tested and cleaned enemy movement as well as adding a damage stat for enemies

+++ ZombosAttack/Player.cpp
@@ -1,4 +1,4 @@
 #include "Player.h"
 
-Player::Player(const float& speed, const unsigned int& health) : 
+Player::Player(const float& speed, const unsigned short& health) : 
 	Character(speed, health){/*Empty*/}
